fix: correctly handle eating potions, time eating using ticks, default to correct drink/eat time from component, send damage event on living entity damage

+++ src/main/java/net/minestom/server/entity/metadata/display/TextDisplayMeta.java
@@ -43,8 +43,8 @@ public class TextDisplayMeta extends AbstractDisplayMeta {
         super.metadata.setIndex(OFFSET + 2, Metadata.VarInt(value));
     }
 
-    public int getTextOpacity() {
-        return super.metadata.getIndex(OFFSET + 3, -1);
+    public byte getTextOpacity() {
+        return super.metadata.getIndex(OFFSET + 3, (byte) -1);
     }
 
     public void setTextOpacity(byte value) {
